Add files via upload
The Image Resizer Tool is a Python code that makes it easy to change the size & type of a lot of images. It uses the Pillow library & works with many image types like JPG, PNG, BMP, & GIF. This tool lets you pick a width & height. It makes sure all images look the same size. It does not change the first files. It can switch image types too. This makes it easy to make image sets look the same. The project is not heavy, simple to use, & shows off file work, error fix & image work with Python.